#375830 - Don't send SIGINT to terminal emulator unnecessarily
Summary: This leaves clearing current command line as default behavior which is sane. However, this change adds a parameter `clearCommand` to the `sendInput()` method to override this behavior. Inserting the file name and pasting are the scenarios where it is actually undesirable to clear the current command.

Test Plan: Inserting file name (Ctrl+Enter) shouldn't clear command line, both when the focus is inside the terminal and on the file list. Same should go for pasting text (Ctrl+V inside the terminal). The automatic `cd` command sent when current directory changes in the list should still clear command line (bug 328068 should not regress).

Reviewers: #krusader, abika, martinkostolny, asensi

Reviewed By: #krusader, abika, martinkostolny, asensi

Subscribers: asensi, martinkostolny, abika, #krusader

Tags: #krusader

Differential Revision: https://phabricator.kde.org/D4419

BUG: 375830

+++ krusader/GUI/terminaldock.cpp
@@ -119,17 +119,19 @@ void TerminalDock::killTerminalEmulator()
     MAIN_VIEW->setTerminalEmulator(false);
 }
 
-void TerminalDock::sendInput(const QString& input)
+void TerminalDock::sendInput(const QString& input, bool clearCommand)
 {
     if (!t)
         return;
 
-    // send SIGINT before input command to avoid unwanted behaviour when current line is not empty
-    // and command is appended to current input (e.g. "rm -rf x " concatenated with 'cd /usr');
-    // code "borrowed" from Dolphin, Copyright (C) 2007-2010 by Peter Penz <peter.penz19@gmail.com>
-    const int processId = t->terminalProcessId();
-    if (processId > 0) {
-        kill(processId, SIGINT);
+    if (clearCommand) {
+        // send SIGINT before input command to avoid unwanted behaviour when current line is not empty
+        // and command is appended to current input (e.g. "rm -rf x " concatenated with 'cd /usr');
+        // code "borrowed" from Dolphin, Copyright (C) 2007-2010 by Peter Penz <peter.penz19@gmail.com>
+        const int processId = t->terminalProcessId();
+        if (processId > 0) {
+            kill(processId, SIGINT);
+        }
     }
 
     t->sendInput(input);
@@ -171,7 +173,7 @@ bool TerminalDock::applyShortcuts(QKeyEvent * ke)
         QString text = QApplication::clipboard()->text();
         if (! text.isEmpty()) {
             text.replace('\n', '\r');
-            sendInput(text);
+            sendInput(text, false);
         }
         return true;
     }
@@ -191,7 +193,7 @@ bool TerminalDock::applyShortcuts(QKeyEvent * ke)
 
         filename = KrServices::quote(filename);
 
-        sendInput(QString(" ") + filename + QString(" "));
+        sendInput(QString(" ") + filename + QString(" "), false);
         return true;
     }
